+++ parser/FnFsets/sets.h
@@ -2,6 +2,6 @@
 #define PARSER_FNFSETS_SETS_H
 
 #include "../../utils/tokens/tokens.h"
-int *firstFollowSet[119];
+short *firstFollowSet[119];
 
 #endif
