+++ utils/errors/errors.h
@@ -1,6 +1,6 @@
 #ifndef UTILS_ERRORS_ERRORS_H
 #define UTILS_ERRORS_ERRORS_H
 
-void lexicalError(char errorChar,int lineNo,char* str);
+void lexicalError(char errorChar,short int lineNo,char* str);
 
 #endif
